requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.12.1
requests==2.31.0
pandas==2.1.3
python-multipart==0.0.6
python-dateutil==2.8.2
aiohttp==3.9.0
apscheduler==3.10.4
email-validator==2.1.0
schedule==1.2.0
